git-svn-id: https://svn.terracotta.org/repo/tc/dso/branches/jdk-15-tests@12332 7fc7bbf3-cf45-46d4-be06-341739edd864

+++ code/base/common-api/src/com/tc/config/schema/setup/StandardL2TVSConfigurationSetupManager.java
@@ -162,7 +162,7 @@ public class StandardL2TVSConfigurationSetupManager extends BaseTVSConfiguration
           if (found) { throw new ConfigurationSetupException("Server{" + serverName
                                                              + "} is part of more than 1 mirror-group:  groups{" + gid
                                                              + "," + groupArray[j].getGroupId() + "}"); }
-          gid = groupArray[j].getGroupId();
+          gid = groupArray[j].getGroupId().toInt();
           found = true;
         }
       }
